Fixed up code gen to use the player
