+++ src/Carnac.Tests/Carnac.Tests.csproj
@@ -55,6 +55,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="KeyPlayer.cs" />
     <Compile Include="SpecificationFor.cs" />
     <Compile Include="ExampleTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
